+++ screen_brightness_platform_interface/lib/constant/method_name.dart
@@ -4,3 +4,6 @@ const methodNameSetScreenBrightness = 'setScreenBrightness';
 const methodNameResetScreenBrightness = 'resetScreenBrightness';
 
 const methodNameHasChanged = 'hasChanged';
+
+const methodNameIsAutoReset = 'isAutoReset';
+const methodNameSetAutoReset = 'setAutoReset';