+++ screen_brightness/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0
+
+* upgraded dependecies
+* changed dependencies version using version range
+* updated README.md
+
+## 0.1.4
+
+* added is auto reset boolean for disable auto reset (iOS only)
+
 ## 0.1.3
 
 * added has changed boolean