+++ screen_brightness_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.0
+
+* fixed code warning
+* changed dependencies version using version range
+
+## 0.0.4
+
+* updated screen_brightness_platform_interface
+
 ## 0.0.3
 
 * added has changed boolean