+++ screen_brightness_ios/CHANGELOG.md
@@ -1,6 +1,19 @@
+## 0.1.0
+
+* added fade animation when brightness changed
+* change dependencies version using version range
+
+## 0.0.5
+
+* added auto reset boolean to prevent auto reset when application lifecycle changed
+
+## 0.0.4
+
+* fixed [#5](https://github.com/aaassseee/screen_brightness/issues/5)
+
 ## 0.0.3
 
-* added has changed bool
+* added has changed boolean
 
 ## 0.0.2